+++ CHANGELOG.md
@@ -2,7 +2,22 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [1.0.2] - 14-11-2020
+## [1.1.0] - 16-11-2020
+### Added
+- Pyup CI + badge
+- jku vulnerability
+- x5u vulnerability
+- new examples
+- new option for cli
+- Makefile
+- new help command
+### Changed
+- new documentation
+- fix print bug([issue](https://github.com/mBouamama/MyJWT/issues/3))
+- Manifest for a better optimization
+- fix examples script
+- new result for facilitate copy / paste
+### [1.0.2] - 14-11-2020
 ### Changed
 - cli
 - unit test